feat(task_delete.js): deletes correct task

+++ controllers/task_controllers/task_delete.js
@@ -1,33 +1,47 @@
+const jwt = require('jsonwebtoken');
+
 const Task = require('../../database/models/TaskModel');
+const User = require('../../database/models/UserModel');
 const {
   clientError: { notFound },
   serverError
 } = require('../../http_status_codes');
 
+const { JWT_SECRET } = process.env;
+
 const taskDelete = async (req, res) => {
   const { id } = req.params;
-  const deletedTask = await Task.findByIdAndDelete(id);
+  const { token } = req.body;
 
-  if (deletedTask === null)
+  try {
+    const { username } = jwt.verify(token, JWT_SECRET);
+    const user = await User.find({ username });
+  
+    if (user) {
+      const deletedTask = await Task.findByIdAndDelete(id);
+      
+      if (deletedTask)
+        return res.json(deletedTask);
+
+      return res.status(notFound).json({
+        status: notFound,
+        statusMessage: 'Could not find the appropriate task to delete'
+      });
+    }
+    
     return res.status(notFound).json({
       status: notFound,
-      statusMessage: 'There is no task by that ID'
+      statusMessage: 'No user found with that name'
     });
+  }
+  catch (err) {
+    console.log(err);
 
-  res.json(deletedTask);
-  
-    // .then(task => {
-    //   if (task === null)
-    //     return res.status(notFound).json({
-    //       status: notFound,
-    //       statusMessage: 'There is no task by that ID'
-    //     })
-      
-    //   res.json(task);
-    // })
-    // .catch(err => {
-    //   res.status(serverError).json(err);
-    // });
+    res.status(serverError).json({
+      status: serverError,
+      statusMessage: 'Server error while trying to delete your task'
+    });
+  }
 };
 
 module.exports = taskDelete;
